yolo26-to-onnx/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
onnx==1.13.1
onnx-simplifier==0.4.36
onnxoptimizer==0.3.13
onnxruntime==1.20.1
onnxslim==0.1.84
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.0
polars==1.38.0
polars-runtime-32==1.38.0
protobuf==3.20.3
psutil==7.2.2
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
scipy==1.15.3
sclblonnx==0.2.1
six==1.17.0
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
triton==3.6.0
typing_extensions==4.15.0
ultralytics==8.4.12
ultralytics-thop==2.0.18
urllib3==2.6.3
